+++ h2gis-functions/src/test/java/org/h2gis/functions/RegressionTest.java
@@ -183,7 +183,6 @@ public class RegressionTest {
         assertTrue(RelateNG.relate(gc0, gc1, RelatePredicate.intersects()));
     }
 
-    @Disabled
     @Test
     public void testBufferPrecision() throws SQLException, ParseException {
         WKTReader wktReader = new WKTReader();