+++ docs/CHANGELOG.md
@@ -5,9 +5,10 @@
 - Write empty geometry with the shapefile driver
 - Do not ask for SRID when geometry is empty
 - Add the name of the file in the FGB header
-- Fix null way id on OSM file- 
+- Fix null way id on OSM file 
 - Use new sonatype auth
 - Add ST_SnapToSelf function
 - Upgrade to H2 2.3.230
 - Upgrade to JTS 1.20
 - Upgrade to H2 2.3.232
+- Add ST_CoverageUnion function